Legg inn tester fra soknad-api
