+++ src/test/resources/application.test.conf
@@ -1,0 +1,59 @@
+no.nav.security.jwt.client.registration {
+  clients = [
+    {
+      client_name = "tokendings"
+      well_known_url = "http://localhost:8080/tokendings/.well-known/openid-configuration"
+      authentication = {
+        client_id = "testClientId"
+        client_jwk = """{"p":"_xCPvqs85ZZVg460Qfot26rQoNRPTOVDo5p4nqH3ep6BK_5TvoU5LFXd26W-1V1Lc5fcvvftClPOT201xgat4DVtliNtoc8od_tWr190A3AzbsAVFOx0nKa5uhLBxP9SsPM84llp6PXF6QTMGFiPYuoLDaQQqL1K4BbHq3ZzF2M","kty":"RSA","q":"7QLqW75zkfSDrn5rMoF50WXyB_ysNx6-2SvaXKGXaOn80IR7QW5vwkleJnsdz_1kr04rJws2p4HBJjUFfSJDi1Dapj7tbIwb0a1szDs6Y2fAa3DlzgXZCkoE2TIrW6UITgs14pI_a7RasclE71FpoZ78XNBvj3NmZugkNLBvRjs","d":"f7aT4poed8uKdcSD95mvbfBdb6X-M86d99su0c390d6gWwYudeilDugH9PMwqUeUhY0tdaRVXr6rDDIKLSE-uEyaYKaramev0cG-J_QWYJU2Lx-4vDGNHAE7gC99o1Ee_LXqMDCBawMYyVcSWx7PxGQfzhSsARsAIbkarO1sg9zsqPS4exSMbK8wyCTPgRbnkB32_UdZSGbdSib1jSYyyoAItZ8oZHiltVsZIlA97kS4AGPtozde043NC7Ik0uEzgB5qJ_tR7vW8MfDrBj6da2NrLh0UH-q28dooBO1vEu0rvKZIescXYk9lk1ZakHhhpZaLykDOGzxCpronzP3_kQ","e":"AQAB","use":"sig","qi":"9kMIR6pEoiwN3M6O0n8bnh6c3KbLMoQQ1j8_Zyir7ZIlmRpWYl6HtK0VnD88zUuNKTrQa7-jfE5uAUa0PubzfRqybACb4S3HIAuSQP00_yCPzCSRrbpGRDFqq-8eWVwI9VdiN4oqkaaWcL1pd54IDcHIbfk-ZtNtZgsOlodeRMo","dp":"VUecSAvI2JpjDRFxg326R2_dQWi6-uLMsq67FY7hx8WnOqZWKaUxcHllLENGguAmkgd8bv1F6-YJXNUO3Z7uE8DJWyGNTkSNK1CFsy0fBOdGywi-A7jrZFT6VBRhZRRY-YDaInPyzUkfWsGX26wAhPnrqCvqxgBEQJhdOh7obDE","alg":"RS256","dq":"7EUfw92T8EhEjUrRKkQQYEK0iGnGdBxePLiOshEUky3PLT8kcBHbr17cUJgjHBiKqofOVNnE3i9nkOMCWcAyfUtY7KmGndL-WIP-FYplpnrjQzgEnuENgEhRlQOCXZWjNcnPKdKJDqF4WAtAgSIznz6SbSQMUoDD8IoyraPFCck","n":"7CU8tTANiN6W_fD9SP1dK2vQvCkf7-nwvBYe5CfANV0_Bb0ZmQb77FVVsl1beJ7EYLz3cJmL8Is1RCHKUK_4ydqihNjEWTyZiQoj1i67pkqk_zRvfQa9raZR4uZbuBxx7dWUoPC6fFH2F_psAlHW0zf90fsLvhB6Aqq3uvO7XXqo8qNl9d_JSG0Rg_2QUYVb0WKmPVbbhgwtkFu0Tyuev-VZ9IzTbbr5wmZwEUVY7YAi73pDJkcZt5r2WjOF_cuIXe-O2vwbOrRgmJfHO9--mVLdATnEyrb6q2oy_75h6JjP-R4-TD1hyoFFoE2gmj-kSS6Z_Gggljs3Aw7--Nh10Q"}"""
+        client_auth_method = private_key_jwt
+      }
+    },
+    {
+      client_name = "azure"
+      well_known_url = "http://localhost:8080/default/.well-known/openid-configuration"
+      authentication = {
+        client_id = "testClientId"
+        client_secret = "mock clientSecret"
+        client_auth_method = client_secret_post
+      }
+    }
+  ]
+}
+
+no.nav.security.jwt {
+    issuers = [
+        {
+             issuer_name = tokendings
+             discoveryurl = "http://localhost:8080/tokendings/.well-known/openid-configuration"
+             accepted_audience = "audience"
+        }
+    ]
+}
+
+endpoints {
+    pdl = ""
+    pdl = ${?PDL_ENDPOINT_URL}
+    pdf = ""
+    pdf = ${?PDF_ENDPOINT_URL}
+    joark = ""
+    joark = ${?JOARK_ENDPOINT_URL}
+    tiltakspengerarena = ""
+    tiltakspengerarena = ${?TILTAKSPENGER_ARENA_ENDPOINT_URL}
+}
+
+audience {
+    pdl = ""
+    pdl = ${?PDL_AUDIENCE}
+    joark = ""
+    joark = ${?JOARK_AUDIENCE}
+    tiltakspengerarena = ""
+    tiltakspengerarena = ${?TILTAKSPENGER_ARENA_AUDIENCE}
+}
+
+scope {
+    pdl = ""
+    pdl = ${?PDL_SCOPE}
+    joark = ""
+    joark = ${?JOARK_SCOPE}
+}
